Beginnings of the Great Restructuring.
