+++ EinsteinEngine/frontend/dimension.py
@@ -15,22 +15,20 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from typing import Union
+from dataclasses import dataclass
 
-from EinsteinEngine.common.stencil_idx import StencilIdxWithCentering, StencilIdxWithNameAndCentering
-
-StencilIdxWithCenteringLike = Union[StencilIdxWithCentering, StencilIdxWithNameAndCentering]
-
-
-def encode_stencil_idx(stencil_idx: StencilIdxWithCenteringLike) -> str:
-    encoded = 'stencil_idx'
-
-    for idx in stencil_idx.indices:
-        if idx >= 0:
-            encoded += f'_{idx}'
-        else:
-            encoded += f'_m{-idx}'
-
-    encoded += f'_{stencil_idx.centering.string_repr}'
-
-    return encoded
+#
+# @dataclass
+# class DimensionSingleton:
+#     value: int = 3
+#
+#
+# _dimension_container = DimensionSingleton()
+#
+#
+# def set_dimension(d: int) -> None:
+#     _dimension_container.value = d
+#
+#
+# def get_dimension() -> int:
+#     return _dimension_container.value
