+++ EinsteinEngine/intermediate/analytic_function_checker.py
@@ -21,18 +21,19 @@ import sympy as sy
 from multimethod import multimethod
 from sympy import Symbol
 
-from EinsteinEngine.dsl.sympywrap import *
+from EinsteinEngine.common.sympywrap import *
 
-x = mkSymbol("x")
-y = mkSymbol("y")
-z = mkSymbol("z")
-t = mkSymbol("t")
+x = mk_symbol("x")
+y = mk_symbol("y")
+z = mk_symbol("z")
+t = mk_symbol("t")
+
 
 class AnalyticFunctionChecker:
-    def __init__(self, params:Set[Symbol], eqns:Dict[Symbol, sy.Expr])->None:
-        self.eqns:Dict[Symbol, sy.Expr] = eqns
-        self.is_analytic:Dict[Symbol, bool] = dict()
-        self.exc:Set[Symbol] = set()
+    def __init__(self, params: Set[Symbol], eqns: Dict[Symbol, sy.Expr]) -> None:
+        self.eqns: Dict[Symbol, sy.Expr] = eqns
+        self.is_analytic: Dict[Symbol, bool] = dict()
+        self.exc: Set[Symbol] = set()
         for k in [x, y, z, t]:
             self.is_analytic[k] = True
             self.exc.add(k)
@@ -40,7 +41,7 @@ class AnalyticFunctionChecker:
             self.is_analytic[k] = True
             self.exc.add(k)
 
-    def analytic(self)->Set[Symbol]:
+    def analytic(self) -> Set[Symbol]:
         done = False
         while not done:
             done = True
@@ -53,21 +54,20 @@ class AnalyticFunctionChecker:
                     done = False
         return set([k for k in self.is_analytic if self.is_analytic[k] == True and k not in self.exc])
 
-
     @multimethod
-    def visit(self, a:sy.Symbol)->Optional[bool]:
+    def visit(self, a: sy.Symbol) -> Optional[bool]:
         return self.is_analytic.get(a, None)
 
     @visit.register
-    def _(self, a:sy.Number)->Optional[bool]:
+    def _(self, a: sy.Number) -> Optional[bool]:
         return True
 
     @visit.register
-    def _(self, a:sy.NumberSymbol)->Optional[bool]:
+    def _(self, a: sy.NumberSymbol) -> Optional[bool]:
         return True
 
     @visit.register
-    def _(self, a:sy.Add|sy.Mul|sy.Function|sy.Pow)->Optional[bool]:
+    def _(self, a: sy.Add | sy.Mul | sy.Function | sy.Pow) -> Optional[bool]:
         for arg in a.args:
             b = self.visit(arg)
             assert b is None or type(b) == bool
@@ -75,19 +75,20 @@ class AnalyticFunctionChecker:
                 return b
         return True
 
+
 if __name__ == "__main__":
-    u = mkSymbol("u")
-    t1 = mkSymbol("t1")
-    p = mkSymbol("p")
+    u = mk_symbol("u")
+    t1 = mk_symbol("t1")
+    p = mk_symbol("p")
 
-    b = AnalyticFunctionChecker({p}, {t1:3+p, u:t1*2+9})
+    b = AnalyticFunctionChecker({p}, {t1: 3 + p, u: t1 * 2 + 9})
     ans = b.analytic()
-    assert ans == {t1,u}, f"ans = {ans}"
+    assert ans == {t1, u}, f"ans = {ans}"
 
-    a = AnalyticFunctionChecker({p}, {t:3+p, u:t*2+9})
+    a = AnalyticFunctionChecker({p}, {t: 3 + p, u: t * 2 + 9})
     assert not a.visit(u)
     assert a.visit(x)
     assert a.visit(sin(x))
-    assert a.visit(sin(x)**2)
-    assert a.visit(sin(x)**2+y)
-    assert not a.visit(sin(u)**2+y)
+    assert a.visit(sin(x) ** 2)
+    assert a.visit(sin(x) ** 2 + y)
+    assert not a.visit(sin(u) ** 2 + y)
