+++ EinsteinEngine/intermediate/coef.py
@@ -20,39 +20,44 @@ from typing import Union
 from multimethod import multimethod
 from sympy import Expr, Add, Mul, Symbol, Idx, Rational, Integer
 
-from EinsteinEngine.dsl.sympywrap import sympify, simplify, mkSymbol, mkIdx
+from EinsteinEngine.common.sympywrap import sympify, simplify, mk_symbol, mk_idx
 
 zero = sympify(0)
-one  = sympify(1)
+one = sympify(1)
+
 
 @multimethod
-def coef(sym:Symbol, expr:Symbol)->Expr:
+def coef(sym: Symbol, expr: Symbol) -> Expr:
     if sym == expr:
         return one
     else:
         return zero
 
+
 @coef.register
-def _(sym:Idx, expr:Idx)->Expr:
+def _(sym: Idx, expr: Idx) -> Expr:
     if sym == expr:
         return one
     else:
         return zero
 
+
 @coef.register
-def _(sym:Union[Idx,Symbol], expr:Union[Integer,Rational])->Expr:
+def _(sym: Union[Idx, Symbol], expr: Union[Integer, Rational]) -> Expr:
     return zero
 
+
 @coef.register
-def _(sym:Union[Symbol,Idx], expr:Add)->Expr:
+def _(sym: Union[Symbol, Idx], expr: Add) -> Expr:
     ret = zero
     for a in expr.args:
         a_ret = coef(sym, a)
         ret += a_ret
     return ret
 
+
 @coef.register
-def _(sym:Union[Symbol,Idx], expr:Mul)->Expr:
+def _(sym: Union[Symbol, Idx], expr: Mul) -> Expr:
     ret = one
     found = False
     for a in expr.args:
@@ -67,25 +72,28 @@ def _(sym:Union[Symbol,Idx], expr:Mul)->Expr:
     else:
         return zero
 
+
 if __name__ == "__main__":
-    a = mkSymbol("a")
-    b = mkIdx("b")
-    c = mkSymbol("c")
+    a = mk_symbol("a")
+    b = mk_idx("b")
+    c = mk_symbol("c")
     one = sympify(1)
 
-    def check(expr:Expr, expected:Expr)->None:
-        print("check:",expr,"->",expected)
+
+    def check(expr: Expr, expected: Expr) -> None:
+        print("check:", expr, "->", expected)
         ret = coef(b, expr)
-        print("    ->",ret)
+        print("    ->", ret)
         should_be_zero = simplify(ret - expected)
         assert zero == should_be_zero, f"{ret} - {expected} == {should_be_zero}"
 
+
     check(b, one)
     check(-b, -one)
-    check(a*c, zero)
-    check(b*a*c, a*c)
-    check(b*c+a*b, a+c)
+    check(a * c, zero)
+    check(b * a * c, a * c)
+    check(b * c + a * b, a + c)
 
-    d = mkIdx("d")
-    check(b-d,one)
-    check(b+d,one)
+    d = mk_idx("d")
+    check(b - d, one)
+    check(b + d, one)
