+++ recipes/Cottonmouth/linear_wave.py
@@ -154,7 +154,7 @@ if __name__ == "__main__":
     Kzz = -Rational(1, 2) * diff(H, t)
 
     # Matrices
-    hij = mkMatrix([
+    hij = mk_matrix([
         [hxx, hxy, hxz],
         [hxy, hyy, hyz],
         [hxz, hyz, hzz],
@@ -166,7 +166,7 @@ if __name__ == "__main__":
         shift_z,
     ]
 
-    Kij = mkMatrix([
+    Kij = mk_matrix([
         [Kxx, Kxy, Kxz],
         [Kxy, Kyy, Kyz],
         [Kxz, Kyz, Kzz],
@@ -181,7 +181,7 @@ if __name__ == "__main__":
         diff(shift_z, t),
     ]
 
-    dt_Kij = mkMatrix([
+    dt_Kij = mk_matrix([
         [diff(Kxx, t), diff(Kxy, t), diff(Kxz, t)],
         [diff(Kxy, t), diff(Kyy, t), diff(Kyz, t)],
         [diff(Kxz, t), diff(Kyz, t), diff(Kzz, t)],
